% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{filter_date}
\alias{filter_date}
\title{filter by date range}
\usage{
filter_date(dt, date_range, DateName = "Date")
}
\arguments{
\item{dt}{data.table or vector}

\item{date_range}{date ranges: c(date_start, date_end)}

\item{DateName}{the name of the Date column in data table}
}
\value{

}
\description{
filter by date range
}
