% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{prints 'Hello, world!'.}
\usage{
hello(pars)
}
\arguments{
\item{pars}{this is a paramenter description}
}
\value{
return the words in the screen
}
\description{
This is a test case for usage of roxygen, anb you can see the descriptions here.
}
\examples{
hello(nothing)
}
