% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_peformance}
\alias{model_peformance}
\title{Model performance}
\usage{
model_peformance(Qsim, Qobs)
}
\arguments{
\item{Qobs}{}
}
\value{

}
\description{
Calculate Statistics include: NSE, logNSE, RMSE, R, R2, Bias, ABias
}
