% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_signatures.R
\name{runoff_signatures}
\alias{runoff_signatures}
\title{Compute runoff signatures}
\usage{
runoff_signatures(Q, Date, Months_sum = 6:8, Months_win = c(12, 1, 2))
}
\arguments{
\item{Months_win}{}
}
\value{

}
\description{
Compute runoff signatures
}
