% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_performance.R
\name{simulation_performance}
\alias{simulation_performance}
\title{Model simulation performance}
\usage{
simulation_performance(Qsim, Qobs)
}
\arguments{
\item{Qobs}{}
}
\value{

}
\description{
Calculate Statistics include: NSE, logNSE, KGE, RMSE, R, R2, Bias, ABias
}
