% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_spatial.R
\name{vis_rasterbrick}
\alias{vis_rasterbrick}
\title{Visualize rasterbrick}
\usage{
vis_rasterbrick(
  rb = NULL,
  bands = NULL,
  cols = NULL,
  ncol = 2,
  outfn = "vis_bands.pdf",
  save = FALSE,
  width = 4,
  height = 4
)
}
\arguments{
\item{rb}{raster brick object}

\item{bands}{numbers of the bands}

\item{cols}{pallettes}

\item{ncol}{number of columns}

\item{outfn}{file name of the output}

\item{save}{indicate export of pdf}

\item{width}{(inchs) of single band image files}

\item{height}{(inchs) of single band image files}
}
\value{

}
\description{
Visualize rasterbrick
}
\examples{
fn <- "I:/projects/fire/victoria/input/GLDAS_bilinear/GLDAS_V21_8d_2000-01-09.tif"
rb <- brick(fn)
names(rb)
cols <- colorRampPalette(RColorBrewer::brewer.pal(11, "Spectral"))
vis_rasterbrick(rb)
vis_rasterbrick(rb, bands = c(1,3,5,7), outfn = "vis_bands2.pdf", cols = cols, save = TRUE)

}
