% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_environment.R
\name{slibrary}
\alias{slibrary}
\title{safe mode for library
check if it is installed before library}
\usage{
slibrary(..., install = TRUE)
}
\arguments{
\item{install}{automatically install the uninstalled packages}

\item{fun}{the name of the packages}
}
\value{

}
\description{
safe mode for library
check if it is installed before library
}
\examples{
slibrary(lattice)
slibrary(lattice, install = FALSE)

}
