% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{slibrary}
\alias{slibrary}
\title{safe mode for library}
\usage{
slibrary(..., install = TRUE)
}
\arguments{
\item{...}{names of packages, same as in library()}

\item{install}{a bool value to indicate whether automatically install
the uninstalled packages, default as TRUE}
}
\value{

}
\description{
library multiple packages with automatic installation
}
\examples{
slibrary(lattice)
slibrary(lattice, sf, raster, install = FALSE)
}
